Mahbub/ComplexDocsRAG/complex_rag.py
import openrouter

# This is a placeholder for a more complex RAG implementation.
# You would typically use a library like LangChain or LlamaIndex
# to handle the retrieval and generation parts.

class ComplexRAG:
    def __init__(self, api_key):
        self.client = openrouter.Client(api_key=api_key)

    def process_documents(self, file_paths):
        # In a real implementation, you would parse these files
        # and create a searchable index.
        print(f"Processing documents: {file_paths}")

    def query(self, question):
        # This is a simplified query to OpenRouter
        response = self.client.chat.completions.create(
            model="google/gemini-pro-1.5", # or another model
            messages=[
                {"role": "user", "content": question}
            ]
        )
        return response.choices[0].message.content

if __name__ == '__main__':
    # Replace with your OpenRouter API key
    rag = ComplexRAG(api_key="YOUR_OPENROUTER_API_KEY")
    rag.process_documents(["contracts/contract.txt", "logs/app.log", "emails/message.eml"])
    answer = rag.query("What is the summary of the documents?")
    print(answer)