Mahbub/requirements.txt
pandas
numpy
sdv
chromadb
langchain
ollama
openrouter-api
speechrecognition
openai-whisper
mindsdb_sdk
streamlit